Fix forever trying to load bad chunk for rendering

+++ src/main/java/atomicedit/backend/chunk/ChunkCoord.java
@@ -101,7 +101,7 @@ public class ChunkCoord {
         return baseAndAdjCoords;
     }
     
-    public int distanceSquaredFrom(int xPos, int zPos) {
+    public long distanceSquaredFrom(long xPos, long zPos) {
         return (xPos - x) * (xPos - x) + (zPos - z) * (zPos - z);
     }
